disable leak sanitizer for now, as it crashes with MB binaries
I am currently not sure whether the crash is due to a memory leak in our
code or bug in leak sanitizer itself. The same binary works correctly under
valgrind. Even weirder - the ASAN:DEADLYSIGNAL is obtained in benign
pthread_create call.

This needs to be investigated further at a later time.